Move RTL compatibility helpers into rtl packages
